+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- Fix type hint for SelectType: only hashable types are allowed.
+- Fix type hint for SelectType: only hashable types are allowed. https://github.com/Textualize/textual/pull/6034
+- Fixed `Content.expand_tabs` https://github.com/Textualize/textual/pull/6038
+- Fixed return value for `Pilot.double_click` and `Pilot.triple_click` https://github.com/Textualize/textual/pull/6035
+
+### Added
+
+- Added `bar_renderable` to `ProgressBar` widget https://github.com/Textualize/textual/pull/5963
 
 # [5.3.0] - 2025-08-07